requirements.txt
sqlalchemy ==2.0
psycopg2-binary ==2.9
pandas ==2.2
requests ==2.32
beautifulsoup4 ==4.12
google-auth ==2.36
pytest-cov ==6.0
google-api-python-client==2.89.03
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
numpy==2.1.3